Move parse and display resvervation Watts to src/common/state_control.[ch].
sview Watts parsing and display was introduced in commit a503a7708e7c4e.
There were differences between scontrol and sview, so this commit moves
the way we parse and display reservation Watts to a common place and make
it so sview and scontrol use these functions. The parse and display
functions have been slightly refactored as compared to the original ones.

Bug 2329